+++ src/Processor/AttachmentProcessor.php
@@ -12,6 +12,7 @@ namespace Elastica\Processor;
 class AttachmentProcessor extends AbstractProcessor
 {
     use Traits\IgnoreMissingTrait;
+    use Traits\TargetFieldTrait;
 
     public const DEFAULT_TARGET_FIELD_VALUE = 'attachment';
     public const DEFAULT_INDEXED_CHARS_VALUE = 100000;
@@ -30,16 +31,6 @@ class AttachmentProcessor extends AbstractProcessor
     public function setField(string $field): self
     {
         return $this->setParam('field', $field);
-    }
-
-    /**
-     * Set target_field. Default attachment.
-     *
-     * @return $this
-     */
-    public function setTargetField(string $targetField): self
-    {
-        return $this->setParam('target_field', $targetField);
     }
 
     /**