+++ src/Processor/BytesProcessor.php
@@ -9,6 +9,7 @@ class BytesProcessor extends AbstractProcessor
 {
     use Traits\IgnoreFailureTrait;
     use Traits\IgnoreMissingTrait;
+    use Traits\TargetFieldTrait;
 
     public function __construct(string $field)
     {
@@ -23,15 +24,5 @@ class BytesProcessor extends AbstractProcessor
     public function setField(string $field): self
     {
         return $this->setParam('field', $field);
-    }
-
-    /**
-     * Set target_field.
-     *
-     * @return $this
-     */
-    public function setTargetField(string $targetField): self
-    {
-        return $this->setParam('target_field', $targetField);
     }
 }