+++ src/Processor/RenameProcessor.php
@@ -13,6 +13,7 @@ class RenameProcessor extends AbstractProcessor
 {
     use Traits\IgnoreFailureTrait;
     use Traits\IgnoreMissingTrait;
+    use Traits\TargetFieldTrait;
 
     public const DEFAULT_IGNORE_MISSING_VALUE = false;
 
@@ -30,15 +31,5 @@ class RenameProcessor extends AbstractProcessor
     public function setField(string $field): self
     {
         return $this->setParam('field', $field);
-    }
-
-    /**
-     * Set target_field.
-     *
-     * @return $this
-     */
-    public function setTargetField(string $targetField): self
-    {
-        return $this->setParam('target_field', $targetField);
     }
 }